python/sphinx_docs/pyproject.toml
[tool.poetry]
name = "opm-python-docs"
version = "0.1.0"
description = """Helper scripts for generating sphinx documentation for the \
                 opm-simulators and opm-common Python bindings"""
authors = ["Håkon Hægland <hakon.hagland@gmail.com>"]
readme = "README.md"
packages = [{ include = "opm_python_docs", from = "src"}]
license = "GPL3"
repository = "https://github.com/OPM/opm-simulators"

[tool.poetry.dependencies]
python = "^3.10"
sphinx = "^7.2.5"
sphinx-rtd-theme = "^1.3.0"
click = "^8.1.7"
sphinx-versioned-docs = "^1.3.1"

[tool.poetry.scripts]
opmdoc-download-files = "opm_python_docs.download_files:main"
opmdoc-view-doc = "opm_python_docs.view_docs:main"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-mock = "^3.14.0"
gitpython = "^3.1.44"
coverage = "^7.6.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
